requirements.txt
Flask==1.1.1
line-bot-sdk==1.15.0
gunicorn==20.0.4
google-cloud-firestore==1.6.1
